exercices/simple-emitter/app/requirements.txt
kafka-python>=1.4.3
flask>=1.0.2